% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_freqpcr.R
\docType{class}
\name{CqFreq-class}
\alias{CqFreq-class}
\alias{CqFreq}
\title{Output object of \code{\link{freqpcr}()}.}
\description{
Output object of \code{\link{freqpcr}()}.
}
\section{Slots}{

\describe{
\item{\code{report}}{A matrix of the simultaneous parameter estimation result. The rows represent the parameters \code{P}, \code{K}, \code{targetScale} (\eqn{\delta_{T}}), \code{sdMeasure} (\eqn{\sigma_{c}}), and \code{EPCR} (\eqn{\eta}).}

\item{\code{obj}}{Returned value of the optimizer function \code{\link{nlm}()} as a list.}

\item{\code{cal.time}}{Calculation time of \code{\link{nlm}()}, stored as a \code{proc_time} class object.}
}}

