% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_freqpcr.R
\name{.freqpcr_loglike}
\alias{.freqpcr_loglike}
\title{Log-likelihood of obtaining Cq values under given parameter set.}
\usage{
.freqpcr_loglike(
  X,
  N,
  DCW,
  DCD,
  zeroAmount,
  para.fixed = NULL,
  beta = TRUE,
  diploid = FALSE
)
}
\arguments{
\item{X}{Numeric vector that stores the parameter values to be optimized via \code{\link{nlm}()}, \code{P} in logit scale and \code{K}, \code{targetScale}, \code{sdMeasure}, and \code{EPCR} in log scale.}

\item{N}{Sample sizes as a numeric vector. \code{N[i]} signifies the number of individuals (both for haploidy and diploidy) contained in the \emph{i}th bulk sample.}

\item{DCW, DCD}{Numeric vectors having the same length as \code{N}. They store the measured values of the two \eqn{\Delta}Cq: \code{DCW (= target0 - housek0)} and \code{DCD (= target1 - housek1)}.}

\item{zeroAmount}{(In RED-\eqn{\Delta\Delta}Cq method) residual rate of restriction enzyme digestion, or (in general \eqn{\Delta\Delta}Cq analyses) small portion of the off-target allele amplified in the PCR process. It needs to be always specified by the user as a number between 0 and 1, usually near 0.}

\item{para.fixed}{Named numeric vector that stores the fixed parameters in \code{\link{freqpcr}()}, if specified by the user. By default (\code{NULL}), all the parameters are considered unknown. Unlike \code{X}, each element should be the raw parameter value in linear scale.}

\item{beta}{Whether to use the beta distribution for the approximation of allele frequency instead of the two gamma distributions for the DNA quantities of the two alleles? Default is \code{TRUE}, which considerably accelerates the calculation.}

\item{diploid}{Is the target organism diploidy? Default is \code{FALSE}, assuming haploidy. Current implementation on the diploids assumes i.i.d. between the amounts of R and S chromosomes owned by a heterozygote, which is unlikely in many animals but necessary for the calculation in a realistic time.}
}
\value{
A scalar of the log likelihood.
}
\description{
The internal function is called from the optimizer (\code{\link[stats]{nlm}()}) running in \code{\link{freqpcr}()}. It defines the log-likelihood by obtaining the two \eqn{\Delta}Cq values (differences in the four Cq measurements) provided that the allele mixing ratio for each bulk sample is given together with other parameters. This function was vectorized over multiple (bulk) samples.
}
