% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_integrator.R
\name{.integrate_beta}
\alias{.integrate_beta}
\title{Integration of likelihood function based on the beta assumption.}
\usage{
.integrate_beta(
  del,
  SHR,
  SHS,
  zeroAmount,
  targetScale,
  sdMeasure,
  xsm = 2,
  EPCR,
  cubmethod = "hcubature",
  relTol = 0.1,
  absTol = 1e-08,
  maxEval = 10^6
)
}
\arguments{
\item{del}{Numeric vector of the observed \eqn{\Delta}Cq values.}

\item{SHR}{The gamma shape parameters for the mutant (R) portion of the bulk samples. Should be the same length as del. Each element of SHR is defined as K*(the assumed number of R allele in the bulk sample).}

\item{SHS}{The gamma shape parameters for the wild (S) portion of the bulk samples. Should be the same length as del. Each element of SHS is defined as K*(the assumed number of S allele in the bulk sample).}

\item{zeroAmount}{(In RED-\eqn{\Delta\Delta}Cq method) residue rate of restriction enzyme digestion, or (in general \eqn{\Delta\Delta}Cq analyses) small portion of the off-target allele on the target locus of the test sample, which will be amplified in the PCR. It needs to be always specified by the user as a number between 0 and 1, usually near 0.}

\item{targetScale}{(\eqn{\delta_{T}}) Scalar. The relative template DNA amount of the target locus to the houskeeping locus. If known, given as a positive numeric.}

\item{sdMeasure}{(\eqn{\sigma_{c}}) Scalar. The measurement error on each Cq value following Normal(0, \eqn{\sigma_{c}^2}). If known, given as a positive numeric.}

\item{xsm}{Specify the accumulation of the standard deviation of the Cq measuring errors when the *-Cq values are fed as difference. For \eqn{\Delta}Cq values, sdMeasure times two. For \eqn{\Delta\Delta}Cq, sdMeasure times four. Default is used in most cases.}

\item{EPCR}{(\eqn{\eta}) Scalar. Amplification efficiency per PCR cycle. If known, given as a positive numeric. When \code{EPCR = 1}, template DNA doubles every cycle (\code{EPCR + 1 = 2}).}

\item{cubmethod}{Cubature method passed to the integrator function. See the section "Methods for cubintegrate".}

\item{relTol}{The maximum tolerance passed to the cubature method. Though the default of cubature::cubintegrate function is 1e-5, the accuracy is reduced here to acceralate the integration.}

\item{absTol}{The absolute tolerance passed to the cubature method. The default is 1e-8, which is less accurate than the default of cubintegrate function (1e-12) but considered enough for the estimation.}

\item{maxEval}{Maximum number of function evaluations needed. The default is 10^6, which is same as the cubintegrate default.}
}
\value{
A numeric vector of marginal likelihoods having the same length as \code{del}.
}
\description{
Internal function to integrate the likelihood getting the \eqn{\Delta}Cq value (the argument \code{del}) over the entire range of the allele ratio (0 to 1). Vectorized for multiple bulk samples. Having the same arguments with .integrate_gamma.
}
\section{Methods for cubintegrate}{
 The following methods are available for \code{\link[cubature]{cubintegrate}()}: cubmethod = c("hcubature", "pcubature", "cuhre", "divonne", "suave", "vegas").
\cr
\code{hcubature} is considerably fast, but less accurate with larger \code{reltol}. \code{cuhre} is moderately fast and the most accurate in most \code{relTol} range. If you can wait longer, \code{hcubature} with \code{relTol = 1e-4} or \code{cuhre} with \code{relTol = 1e-1} is recommended.
At \code{reltol = 1e-1}, \code{hcubature} is three times faster than \code{cuhre}, but the log-likelihood fluctuate by 0.1 (by 0.001 in \code{cuhre}).
The speed and accuracy of \code{hcubature} at \code{reltol = 1e-5} is comparable with \code{cuhre} at \code{reltol = 1e-1}.
\cr
\code{pcubature} and \code{divonne} frequently returns \code{NaN} and are not recommended.
\cr
\code{suave} and \code{vegas} are Monte Carlo integration and slow. They are certainly accurate even at \code{reltol = 1}, but interior to \code{cuhre} with same \code{reltol}.
}

\seealso{
Other integrators: 
\code{\link{.integrate_gamma}()}
}
\concept{integrators}
\keyword{internal}
