% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{.knownqpcr_loglike}
\alias{.knownqpcr_loglike}
\title{Log-likelihood of getting Cq values when exact allele mixing ratios are known.}
\usage{
.knownqpcr_loglike(X, A, trueY, Digest, Gene, Cq)
}
\arguments{
\item{X}{A numeric vector that stores the current parameter sizes of \code{meanDNA}, \code{targetScale}, \code{baseChange}, \code{sdMeasure}, \code{zeroAmount}, and \code{EPCR} in log scale.}

\item{A}{Optionally, you can specify relative DNA content between the samples, as a numeric vector having the same length as the Cq data vectors. It is the counterpart of the \code{N} argument in \code{\link{freqpcr}()}, but an element of \code{A} is not restricted to integer. Because the concentration as a whole is also adjusted with the parameter \code{meanDNA} (see Value section), this variable should be used exclusively to reflect the relative DNA contents between the samples solutions. Otherwise, it is better to keep its default setting (1 for all replicates).}

\item{trueY}{A numeric vector having the same length as the Cq data vectors. \code{trueY[i]} signifies the exact frequency of the mutant allele in the \emph{i}th sample. The values must be between 0 and 1. To improve the estimation accuracy, y == 1 (pure mutant solution) should be added to your experimental design.}
}
\value{
A scalar of the log likelihood.
}
\description{
Internal function to return the log-likelihood getting the four Cq measurements under true allele frequency for each bulk sample is known.
}
\keyword{internal}
