% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{.knownqpcr_loglike_paired}
\alias{.knownqpcr_loglike_paired}
\title{Log-likelihood of obtaining Cq values when exact allele content is known.}
\usage{
.knownqpcr_loglike_paired(X, A, trueY, housek0, target0, housek1, target1)
}
\arguments{
\item{X}{A numeric vector that stores the current parameter sizes of \code{meanDNA}, \code{targetScale}, \code{baseChange}, \code{sdMeasure}, \code{zeroAmount}, and \code{EPCR} in log scale.}

\item{A}{Optionally, you can specify relative DNA content between the samples, as a numeric vector having the same length as the Cq data vectors. It is the counterpart of the \code{N} or \code{A} argument in \code{\link{freqpcr}()}, whereas an element of \code{A} is not restricted to integer. Because the concentration as a whole is also adjusted with the parameter \code{meanDNA} (see Value section), this variable should be used exclusively to reflect the relative DNA contents between the samples solutions. Otherwise, it is better to keep its default setting (1 for all replicates).}

\item{trueY}{A numeric vector having the same length as the Cq data vectors. \code{trueY[i]} signifies the exact frequency of the mutant allele in the \emph{i}th sample. The values must be between 0 and 1. To improve the estimation accuracy, y == 1 (pure mutant solution) should be included in your experimental design.}

\item{target0, target1, housek0, housek1}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}
}
\value{
A scalar of the log likelihood.
}
\description{
The internal function to calculate the log-likelihood by obtaining the sets of four Cq measurements under true allele frequency for each bulk sample is known. As the largest difference from \code{\link{.freqpcr_loglike}()}, this function does not take account for the uncertainty of the individual DNA yield.
}
\keyword{internal}
