% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{.knownqpcr_loglike_unpaired}
\alias{.knownqpcr_loglike_unpaired}
\title{Log-likelihood of getting Cq values when exact allele content is known.}
\usage{
.knownqpcr_loglike_unpaired(X, A, trueY, Gene, Digest, Cq)
}
\arguments{
\item{X}{A numeric vector that stores the current parameter sizes of \code{meanDNA}, \code{targetScale}, \code{baseChange}, \code{sdMeasure}, \code{zeroAmount}, and \code{EPCR} in log scale.}

\item{A}{Sample sizes as a numeric vector.}

\item{trueY}{A numeric vector of the same length as \code{A}. \code{trueY[i]} signifies the true frequency of the mutant allele in the \emph{i}th bulk sample. The values must be between 0 and 1.}

\item{Gene}{Numeric vector that specify the gene is housekeeping (= 0) or target (= 1), having the same length as \code{A} and \code{trueY}.}

\item{Digest}{Numeric vector having the same length as \code{A} and \code{trueY}. It specify the gene is intact (= 0) or digested with endonuclease (= 1) in the RED-\eqn{\Delta\Delta}Cq method. In general \eqn{\Delta\Delta}Cq analyses, the argument means all the alleles on the target locs will be amplified (= 0) (it has ) or only the R portion will be amplified (using an R-specific primer set).}

\item{Cq}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}
}
\value{
A scalar of the log likelihood.
}
\description{
Internal function to return the log-likelihood getting the four Cq measurements under true allele frequency for each bulk sample is known.
}
\keyword{internal}
