% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{knownqpcr_loglike}
\alias{knownqpcr_loglike}
\title{Log-likelihood of getting Cq values when exact allele content is known.}
\usage{
knownqpcr_loglike(X, A, trueY, target0, target1, housek0, housek1)
}
\arguments{
\item{X}{A numeric vector that stores the current parameter sizes of \code{meanDNA}, \code{targetScale}, \code{baseChange}, \code{sdMeasure}, \code{zeroAmount}, and \code{EPCR} in log scale.}

\item{A}{Sample sizes as a numeric vector. It is the counterpart of the \code{N} argument in \code{\link{freqpcr}()}, but the elements of \code{A} in \code{\link{knownqpcr}()} are not restricted to integer. Including the case you have arranged the artificial sample solution with arbitrary DNA concentration, it is convenient to determine \code{A} in order to ensure that the equivalent amount of DNA extracted from an individual is approximately 1, though it is not a prerequisite and you virtually can put any positive numeric.}

\item{trueY}{A numeric vector of the same length as \code{A}. The size of \code{trueY[i]} signifies the true frequency of the mutant allele in the \emph{i}th bulk sample. The values must be between 0 and 1.}

\item{target0, target1, housek0, housek1}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}
}
\value{
A scalar of the log likelihood.
}
\description{
Internal function to return the log-likelihood getting the four Cq measurements under true allele frequency for each bulk sample is known. As the largest difference from \code{\link{freqpcr_loglike}()}, this function does not take account for the uncertainty in individual DNA yield.
}
\seealso{
Other likelihood definitions: 
\code{\link{freqpcr_loglike}()}
}
\concept{likelihood definitions}
