% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{.knownqpcr_loglike_paired}
\alias{.knownqpcr_loglike_paired}
\title{Log-likelihood of obtaining Cq values when exact allele content is known.}
\usage{
.knownqpcr_loglike_paired(X, A, trueY, housek0, target0, housek1, target1)
}
\arguments{
\item{X}{A numeric vector that stores the current parameter sizes of \code{meanDNA}, \code{targetScale}, \code{baseChange}, \code{sdMeasure}, \code{zeroAmount}, and \code{EPCR} in log scale.}

\item{A}{Sample sizes as a numeric vector. It is the counterpart of the \code{N} argument in \code{\link{freqpcr}()}, but an element of \code{A} is not restricted to integer. Considering the case you have arranged the artificial sample solution with arbitrary DNA concentration, it is convenient to determine \code{A} to ensure that the equivalent amount of DNA extracted from an individual is approximately 1. As the concentration in total is also adjusted with the parameter \code{targetScale} (see Value section), this variable should be used exclusively to reflect the differences in relative concentration between samples.}

\item{trueY}{A numeric vector of the same length as \code{A}. \code{trueY[i]} signifies the true frequency of the mutant allele in the \emph{i}th bulk sample. The values must be between 0 and 1.}

\item{target0, target1, housek0, housek1}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}
}
\value{
A scalar of the log likelihood.
}
\description{
The internal function to calculate the log-likelihood by obtaining the sets of four Cq measurements under true allele frequency for each bulk sample is known. As the largest difference from \code{\link{.freqpcr_loglike}()}, this function does not take account for the uncertainty of the individual DNA yield.
}
\keyword{internal}
