% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_package.R
\docType{package}
\encoding{UTF-8}
\name{freqpcr-package}
\alias{freqpcr-package}
\alias{_PACKAGE}
\title{The freqpcr package}
\description{
Maximum likelihood estimation of allele frequency from quantitive PCR based on the double-delta Cq method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sudoms/freqpcr}
}

}
\author{
\strong{Maintainer}: Masaaki Sudo \email{masaaki@sudori.info} (\href{https://orcid.org/0000-0001-9834-9857}{ORCID})

}
\keyword{internal}
