% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_freqpcr.R
\name{freqpcr}
\alias{freqpcr}
\title{Estimate population allele frequency from the set of Cq measurements.}
\usage{
freqpcr(
  N,
  housek0,
  target0,
  housek1,
  target1,
  P = NULL,
  K = NULL,
  targetScale = NULL,
  sdMeasure = NULL,
  EPCR = 0.99,
  XInit0 = c(P = NULL, K = NULL, targetScale = NULL, sdMeasure = NULL, EPCR = NULL),
  zeroAmount = 0.0016,
  beta = TRUE,
  diploid = FALSE,
  pvalue = 0.05,
  gradtol = 1e-04,
  steptol = 1e-09,
  iterlim = 100,
  maxtime = 600,
  print.level = 1,
  ...
)
}
\arguments{
\item{N}{Sample sizes as a numeric vector. \code{N[i]} signifies the number of individuals (both for haploidy and diploidy) contained in the \emph{i}th bulk sample.}

\item{housek0}{A numeric vector having the same length as \code{N}. In RED-\eqn{\Delta\Delta}Cq method, \code{housek0} is the Cq values of the test sample without the restriction enzyme digestion, which is amplified with the primer set for a housekeeping gene. In a general \eqn{\Delta\Delta}Cq analyses, \code{housek0} is defined for the control sample (100\% mutant) solution, which is also amplified with the primer set for the housekeeping gene.}

\item{target0}{A numeric vector having the same length as \code{N}. In RED-\eqn{\Delta\Delta}Cq method, \code{target0[i]} signifies the measured Cq value of the \emph{i}th bulk sample without the digestion, which is treated with the primer set that amplifies both alleles, wild-type (or S: susceptible) and mutant (or R: resistant to a pesticide), on the target locus. In general \eqn{\Delta\Delta}Cq analyses, \code{target0} is the Cq values of the control sample, which is defined as a sample solution that contains 100\% R, amplified with an R-specific primer set.}

\item{housek1}{A numeric vector having the same length as \code{N}. The Cq values of the test sample measured on the housekeeping gene after the restriction enzyme digestion (in RED-\eqn{\Delta\Delta}Cq method), or the test sample amplified on the housekeeping gene (in general \eqn{\Delta\Delta}Cq analyses).}

\item{target1}{A numeric vector having the same length as \code{N}. For each test sample with unknown allele-ratio, \code{target1[i]} is defined as the Cq value for the target locus amplified with the universal primer set after the restriction enzyme digestion (in RED-\eqn{\Delta\Delta}Cq method), or the target locus amplified with the R-allele-specific primer set (in general \eqn{\Delta\Delta}Cq analyses).}

\item{P}{Population allele frequency from which the test samples are retrieved. Default is \code{NULL} and to be estimated. If the parameter is known, it should be given as a numeric between 0 and 1.}

\item{K}{The gamma shape parameter of the individual DNA yield. Default is \code{NULL} and to be estimated. If known, given as a positive numeric.}

\item{targetScale}{(\eqn{\delta_{T}}) The relative template DNA amount of the target locus to the houskeeping locus. If known, given as a positive numeric.}

\item{sdMeasure}{(\eqn{\sigma_{c}}) The measurement error on each Cq value following Normal(0, \eqn{\sigma_{c}^2}). If known, given as a positive numeric.}

\item{EPCR}{(\eqn{\eta}) Amplification efficiency per PCR cycle. If known, given as a positive numeric. When \code{EPCR = 1}, template DNA doubles every cycle (\code{EPCR + 1 = 2}).}

\item{XInit0}{Optionally the initial value for the parameter optimization can be specified, but the user is highly recommended to keep the argument as is. Unlike XInit in \code{\link{knownqpcr}()}, the argument is not directly fed to the optimizer. Each value is used only when the parameter is unknown.}

\item{zeroAmount}{(In RED-\eqn{\Delta\Delta}Cq method) residual rate of restriction enzyme digestion, or (in general \eqn{\Delta\Delta}Cq analyses) small portion of the off-target allele amplified in the PCR process. It needs to be always specified by the user as a number between 0 and 1, usually near 0.}

\item{beta}{Whether to use the beta distribution for the approximation of allele frequency instead of the two gamma distributions for the DNA quantities of the two alleles? Default is \code{TRUE}, which considerably accelerates the calculation.}

\item{diploid}{Is the target organism diploidy? Default is \code{FALSE}, assuming haploidy. Current implementation on the diploids assumes i.i.d. between the amounts of R and S chromosomes owned by a heterozygote, which is unlikely in many animals but necessary for the calculation in a realistic time.}

\item{pvalue}{The two-sided confidence interval is calculated at the last iteration based on the significance level. Default is 0.05, which returns the 95\% CI (2.5 to 97.5 percentile) based on the Hessian matrix.}

\item{gradtol, steptol, iterlim}{Arguments passed to \code{\link[stats]{nlm}()}. \code{gradtol} and \code{steptol} are the positive scalars giving the tolerance to terminate the algorithm and the minimum allowable relative step length. \code{iterlim} specifies the maximum number of iterations to be performed before the program is terminated. Usually 30 iterations are enough.}

\item{maxtime}{a positive scalar to set the maximum calculation time in seconds to abort the optimizer (and return error). The total calculation time largely depends on \code{N[i]}, the number of individuals contained in each bulk sample.}

\item{print.level}{\code{print.level=1} (the default) shows the initial values of the parameters and likelihood as well as the output in the last iteration. \code{print.level=2} shows the parameter values and gradients in every step. \code{print.level=0} does not output any intermediate state to R console, simply returning the result summary.}

\item{...}{Additional arguments passed to the function.}
}
\value{
Object of the S4 class \linkS4class{CqFreq}. The slot \code{report} is a matrix and each row contains the estimated parameter value with 100*(1-pvalue)\% confidence interval. The following parameters are returned:
\enumerate{
    \item\code{P} Population allele frequency from which the test samples are retrieved.
    \item\code{K} The gamma shape parameter of the individual DNA yield.
    \item\code{targetScale} (\eqn{\delta_{T}}), the relative template DNA amount of the target to the houskeeping loci.
    \item\code{EPCR} (\eqn{\eta}), the amplification efficiency per PCR cycle.
    \item\code{sdMeasure} or "Cq measurement error" (\eqn{\sigma_{c}}).
}
}
\description{
The function estimates the population allele frequency using the dataset of Cq values measured over \code{length(N)} bulk samples, each of which has a sample size of \code{N[i]} as the number of individuals included. \code{N[i]} can be 1, meaning that every individual is analyzed separately. For the \emph{i}th (bulk) sample, the four Cq values were measured as \code{target0[i]}, \code{target1[i]}, \code{housek0[i]}, and \code{housek1[i]}, respectively.
\cr
The function can estimate up to five parameters simultaneously when the Cq measures are available for more than two bulk samples: \code{P}, \code{K}, \code{targetScale}, \code{sdMeasure}, and \code{EPCR}, although \code{EPCR} is not reliable in many cases and users are encouraged to know the size of \code{EPCR} beforehand by using \code{\link{knownqpcr}()} or \code{\link{knownqpcr_unpaired}()}. By default, parameters other than the \code{EPCR} were estimated.
}
\section{Choise of the parameters to be estimated}{

Estimation is conducted only for parameters specified, explicitly or implicitly, as \code{NULL}. If one feeds a value e.g. \code{K=1} or \code{sdMeasure=0.24}, it is then treated as fixed parameter.
\cr
\code{EPCR} is the amplification efficiency of PCR and usually not estimable from the experiments with unknown allele ratios. You should verify the size of \code{EPCR} and \code{zeroAmount} (the residue rate of the restriction enzyme digestion in the RED-\eqn{\Delta\Delta}Cq method) prior to the experiment with unknown samples. The functions \code{\link{knownqpcr}()} and \code{\link{knownqpcr_unpaired}()} provide the procedure to estimate the sizes of the experimental parameters using the DNA solutions of known allele mixing ratios.
\cr
For the unknown parameters, \code{XInit0} optionally specifies initial values for the optimization using \code{\link[stats]{nlm}()} though the use of internal default is highly recommended. The specification as a fixed parameter has higher priority than the specification in \code{XInit0}. Every user-specified parameter values, fixed or unknown, must be given in linear scale (e.g. between 0 and 1 for the allele frequency); they are transformed internally to log or logit.
}

\examples{
# Dummy Cq dataset with six bulk samples, each of which comprises of eight haploids.
EPCR <- 0.95; zeroAmount <- 0.0016; # True values for the two must be known.
P <- 0.75
dmy_cq <- make_dummy(   rand.seed=1, P=P, K=2, ntrap=6, npertrap=8,
                        scaleDNA=1e-07, targetScale=1.5, baseChange=0.3,
                        EPCR=EPCR, zeroAmount=zeroAmount,
                        sdMeasure=0.3, diploid=FALSE   )
print(dmy_cq)
# Estimation with freqpcr, where P, K, targetScale, and baseChange are marked unknown.
result <- freqpcr( dmy_cq@N,
                   dmy_cq@housek0, dmy_cq@target0, dmy_cq@housek1, dmy_cq@target1,
                   EPCR=EPCR, zeroAmount=zeroAmount, beta=TRUE, print.level=2 )
print(result)
# Estimation with freqpcr, where you have knowledge on the size of baseChange.
result <- freqpcr( dmy_cq@N,
                   dmy_cq@housek0, dmy_cq@target0, dmy_cq@housek1, dmy_cq@target1,
                   baseChange=0.3,
                   EPCR=EPCR, zeroAmount=zeroAmount, beta=TRUE, print.level=1 )
}
\seealso{
Other estimation procedures: 
\code{\link{knownqpcr_unpaired}()},
\code{\link{knownqpcr}()},
\code{\link{sim_dummy}()}
}
\concept{estimation procedures}
