% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{knownqpcr}
\alias{knownqpcr}
\title{Estimate auxiliary parameters using the bulk samples with known allele ratios.}
\usage{
knownqpcr(
  A,
  trueY,
  target0,
  target1,
  housek0,
  housek1,
  XInit = c(meanDNA = -10, targetScale = 0, baseChange = 0, sdMeasure = 1, zeroAmount =
    -5, EPCR = 0),
  method = "BFGS",
  pvalue = 0.05,
  trace = 0,
  report = 10,
  quiet = FALSE
)
}
\arguments{
\item{A}{Sample sizes as a numeric vector. It is the counterpart of the \code{N} argument in \code{\link{freqpcr}()}, but the elements of \code{A} in \code{\link{knownqpcr}()} are not restricted to integer. Including the case you have arranged the artificial sample solution with arbitrary DNA concentration, it is convenient to determine \code{A} in order to ensure that the equivalent amount of DNA extracted from an individual is approximately 1, though it is not a prerequisite and you virtually can put any positive numeric.}

\item{trueY}{A numeric vector of the same length as \code{A}. The size of \code{trueY[i]} signifies the true frequency of the mutant allele in the \emph{i}th bulk sample. The values must be between 0 and 1.}

\item{target0, target1, housek0, housek1}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}

\item{XInit}{A named vector specifying the initial sizes of the auxiliary parameters in the optimization. Defined in the natural log scale e.g. \code{zeroAmount = -5} corresponds to the residue rate of \code{exp(-5)} = 0.007. It is highly recommended to keep the default value.}

\item{method}{A string specifying the optimization algorithm used in \code{\link[stats]{optim}()}. The default is \code{BFGS}, which is plausible in most situation.}

\item{pvalue}{The two-sided confidence interval is calculated at the last iteration based on the significance level. Default is 0.05, which returns the 95\% CI (2.5 to 97.5 percentile) based on the Hessian matrix.}

\item{trace}{Non-negative integer. If positive, \code{\link[stats]{optim}()} outputs trace information. The default is 0 (no information).}

\item{report}{The frequency of reports if \code{trace} is positive. Defaults to every 10 iterations.}

\item{quiet}{Suppress the output of the function sent to stdout? The default is \code{FALSE}.}
}
\value{
A table containing the estimated values for the following parameters:
\cr
\code{meanDNA} is the template DNA concentration.
\cr
\code{targetScale} (\eqn{\delta_{T}}) is the relative template DNA amount of the target to the houskeeping loci.
\cr
\code{baseChange} (\eqn{\delta_{B}}) is the change rate in the amount of the internal reference after digestion (in the RED-delta delta Cq method).
\cr
\code{sdMeasure} (\eqn{\sigma_{c}}) is the measurement error on each of the four Cq values.
\cr
\code{EPCR} (\eqn{\eta}) is the amplification efficiency per PCR cycle.
}
\description{
Function to estimate the auxiliary experimental parameters using the sample solutions where the exact allele ratios (the argument \code{trueY}) are known.
}
\examples{
# First, make a dummy Cq dataset of six bulk DNA samples,
# each of which comprises eight haploid individuals.
ntrap <- 6
npertrap <- 8
dmy_cq <- makeCqList(   rand.seed=1, P=0.75, K=2, ntrap=ntrap, npertrap=npertrap,
                        scaleDNA=1e-07, targetScale=1.5, baseChange=0.3, EPCR=0.95,
                        zeroAmount=1e-3, sdMeasure=0.3, diploid=FALSE   )
print(dmy_cq)

# Then, estimate the auxiliary parameters.
# Note that "dmy_cq" does not include information on the parameters.
knownqpcr(  dmy_cq@N, rep(0.75, ntrap),
            dmy_cq@target0, dmy_cq@target1, dmy_cq@housek0, dmy_cq@housek1,
            quiet=TRUE  )
}
\seealso{
Other estimation procedures: 
\code{\link{freqpcr}()},
\code{\link{simCqList}()}
}
\concept{estimation procedures}
