% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_known_ratio.R
\name{knownqpcr_unpaired}
\alias{knownqpcr_unpaired}
\title{Estimate auxiliary parameters when the sample pairs are incomplete.}
\usage{
knownqpcr_unpaired(
  A,
  trueY,
  Gene,
  Digest,
  Cq,
  XInit = c(meanDNA = -10, targetScale = 0, baseChange = 0, sdMeasure = 1, zeroAmount =
    -5, EPCR = 0),
  method = "BFGS",
  pvalue = 0.05,
  trace = 0,
  report = 10,
  quiet = FALSE
)
}
\arguments{
\item{A}{Sample sizes as a numeric vector.}

\item{trueY}{A numeric vector of the same length as \code{A}. \code{trueY[i]} signifies the true frequency of the mutant allele in the \emph{i}th bulk sample. The values must be between 0 and 1.}

\item{Gene}{Numeric vector that specify the gene is housekeeping (= 0) or target (= 1), having the same length as \code{A} and \code{trueY}.}

\item{Digest}{Numeric vector having the same length as \code{A} and \code{trueY}. It specify the gene is intact (= 0) or digested with endonuclease (= 1) in the RED-\eqn{\Delta\Delta}Cq method. In general \eqn{\Delta\Delta}Cq analyses, the argument means all the alleles on the target locs will be amplified (= 0) (it has ) or only the R portion will be amplified (using an R-specific primer set).}

\item{Cq}{Measured Cq values. Numeric vectors having the same length as \code{A} and \code{trueY}.}

\item{XInit}{A named vector specifying the initial sizes of the auxiliary parameters in the optimization. Defined in the natural log scale; e.g. \code{zeroAmount = -5} corresponds to the residue rate of \code{exp(-5)} = 0.007. It is highly recommended to keep the default values.}

\item{method}{A string specifying the optimization algorithm used in \code{\link[stats]{optim}()}. The default is \code{BFGS}, which is plausible in most situation.}

\item{pvalue}{The two-sided confidence interval is calculated at the last iteration based on the significance level. Default is 0.05, which returns the 95\% CI (2.5 to 97.5 percentile) based on the Hessian matrix.}

\item{trace}{Non-negative integer. If positive, \code{\link[stats]{optim}()} outputs trace information. The default is 0 (no information).}

\item{report}{The frequency of reports if \code{trace} is positive. Defaults to every 10 iterations.}

\item{quiet}{Suppress the output of the function sent to stdout? The default is \code{FALSE}.}
}
\value{
A table containing the estimated parameter values, which is same as \code{\link{knownqpcr}()}.
}
\description{
A variant of \code{\link{knownqpcr}()} to estimate the auxiliary parameters when the exact mixing ratios are known for each the sample, but the qPCR analyses are conducted for only a part of the sample sets. To deal with the cases that the combination of the four Cq measurements is incomplete (thus the data lengths of \code{housek0}, \code{target0}, \code{housek1}, and \code{target1} are different), the function accepts the observed Cq values concatenated into a single vector (the argument \code{Cq}) accompanied with the experimental conditions (the arguments \code{Gene} and \code{Digest}).
}
\examples{
# A dummy Cq dataset: four mixing ratios (0.1, 0.25, 0.5, 1) with four replicates.
# Settings: targetScale=1.5, baseChange=0.3, EPCR=0.95, zeroAmount=1e-3, sdMeasure=0.3

trueY <- c(rep(0.1, 4), rep(0.25, 4), rep(0.5, 4), rep(1, 4))
housek0 <- c( 19.39, 19.78, 19.28, 19.58,  18.95, 19.91, 19.66, 19.96,
              20.05, 19.86, 19.55, 19.61,  19.86, 19.27, 19.59, 20.21 )
target0 <- c( 19.16, 19.08, 19.28, 19.03,  19.17, 19.67, 18.68, 19.52,
              18.92, 18.79, 18.8, 19.28,   19.57, 19.21, 19.05, 19.15 )
housek1 <- c( 21.61, 21.78, 21.25, 21.07,  22.04, 21.45, 20.72, 21.6,
              21.51, 21.27, 21.08, 21.7,   21.44, 21.46, 21.5, 21.8 )
target1 <- c( 24.3, 24.22, 24.13, 24.13,   22.74, 23.14, 23.02, 23.14,
              21.65, 22.62, 22.28, 21.65,  20.83, 20.82, 20.76, 21.3 )

# Complete observation: "wide" format.
d.cmp <- data.frame(A=rep(1, 16), trueY, housek0, target0, housek1, target1)
print(d.cmp)
p.cmp <- knownqpcr( d.cmp$A, d.cmp$trueY,
                    d.cmp$housek0, d.cmp$target0, d.cmp$housek1, d.cmp$target1 )
print(p.cmp)

# Incomplete observation: prepared in the "long" format.
# Undegested (housek0 and target0) samples are analyzed only for trueY == 1.
d.long <- data.frame(
    A=rep(1, 4+4+16+16),
    Gene=c(rep(0, 4), rep(1, 4), rep(0, 16), rep(1, 16)),
    Digest=c(rep(0, 4+4), rep(1, 16+16)),
    trueY=c(rep(1, 8), rep(c(rep(0.1, 4), rep(0.25, 4), rep(0.5, 4), rep(1, 4)), 2)),
    Repl=rep(1:4, 1+1+4+4),
    Cq=c(housek0[13:16], target0[13:16], housek1, target1))
print(d.long)
p.long <- knownqpcr_unpaired( d.long$A, d.long$trueY,
                              d.long$Gene, d.long$Digest, d.long$Cq )
print(p.long)
}
\seealso{
Other estimation procedures: 
\code{\link{freqpcr}()},
\code{\link{knownqpcr}()},
\code{\link{sim_dummy}()}
}
\concept{estimation procedures}
